% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karma_adf.r
\name{karma.adf}
\alias{karma.adf}
\title{Augmented Dickey-Fuller test for stationarity.}
\usage{
karma.adf(y, modeltype = "none", lags = 1, diffs = 0, log = F, dw = F,
  stdout = T)
}
\arguments{
\item{y}{A univariate time-series vector; type <numeric> or <ts>.}

\item{modeltype}{Set ADF model with intercept and trend, intercept only, or neither. Takes values: "trend", "drift", "none".}

\item{lags}{Length of lags for autoregression.}

\item{diffs}{Differencing step. Indicate whether the input series needs to be differenced for stationarity (and to what degree); {0,1,...,n}; type <int>.}

\item{log}{Logarithmic transformation flag. Indicate whether the input series needs to be log-transformed for stationarity; {T, F}; type <logical>.}

\item{stdout}{Option to print out all test diagnostics; <logical>.}
}
\value{
Object of class "karma.adf".
}
\description{
Augmented Dickey-Fuller test for stationarity.
}
\examples{
# Apply ADF test and print out diagnostics:
adf.object <- karma.adf(WWWusage, modeltype = "drift", lags = 1, diffs = 0, log = F, stdout = T)
print(adf.object$stationarity)
}
\seealso{
\code{\link{tseries}}, \code{\link{forecast}}, \code{\link{car}}
}

