% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karma_boxjenkins.r
\name{karma.boxjenkins}
\alias{karma.boxjenkins}
\title{Box-Jenkins method for ARMA model selection on a stationary process.}
\usage{
karma.boxjenkins(y, diffs = 0, log = F, fixed = F, xreg = NULL,
  N = 100, box_test = F, max_ar = 20, max_ma = 20, max_conv = 2,
  max_iter = 200, plot = F, stdout = T)
}
\arguments{
\item{y}{A univariate time-series vector; type <numeric> or <ts>.}

\item{diffs}{Differencing step: Indicates whether the input series needs to be differenced for stationarity (and to what degree); {0,1,...,n}; type <int>.}

\item{log}{Logarithmic transformation flag. Indicates whether the input series needs to be log-transformed for stationarity; {T, F}; type <logical>.}

\item{fixed}{Fixed term flag. Indicates whether the fixed term option in Arima() needs to be switched on during model selection; {T, F}; type <logical>.}

\item{xreg}{Optional vector or matrix of exogenous regressors; see documentation for Arima(), package 'forecast'.}

\item{N}{Maximum lag at which to calculate autocorrelation and partial autocorrelatin functions; see documentation for acf(), pacf().}

\item{box_test}{T/F flag. Indicates whether or not a Box-Pierce test for autocorrelation should be performed at every algorithm iteration.}

\item{max_ar}{Maximum AR term (value of p).}

\item{max_ma}{Maximum MA term (value of q).}

\item{max_conv}{Maximum number of iterations without improvement before the algorithm converges forcefully (stuck to a local optimum).}

\item{max_iter}{Maximum number of iterations without improvement before the algorithm converges naturally (reached a global or local optimum).}

\item{plot}{Option to depict plots during local search; if TRUE (default), AC and PAC plots are active. <logical>}

\item{stdout}{Option to output optimisation diagnostics during local search; <logical>}
}
\value{
Object of class "karma.fit"; (extends class "Arima" from package 'forecast').
}
\description{
Box-Jenkins method for ARMA model selection on a stationary process.
}
\examples{
# Find transformation steps required for stationarity:
stationarity.options <- karma.transform(ldeaths, stdout = F, autolog = F, autodiffs = 1)
# Apply Box-Jenkins method on the stationary series:
boxj.fit <- karma.boxjenkins(ldeaths, diffs = stationarity.options$diffs, log = stationarity.options$log)
# Apply cross-validation and calculate MAPE on out-of-sample (test) data:
karma.cv(boxj.fit)
}
\seealso{
\code{\link{tseries}}, \code{\link{forecast}}
}

