% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karma_cast.r
\name{karma.cast}
\alias{karma.cast}
\title{Convert "ARIMA" object to "karma.fit" object.}
\usage{
karma.cast(arma0)
}
\arguments{
\item{arma0}{"ARIMA" object.}
}
\value{
Object of class "karma.fit"; (extends class "ARIMA" from package 'forecast').
}
\description{
Convert "ARIMA" object to "karma.fit" object.
}
\examples{
karma.cv( karma.cast(auto.arima(WWWusage)) )
}
\seealso{
\code{\link{tseries}}, \code{\link{forecast}}
}

