% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karma_cv.r
\name{karma.cv}
\alias{karma.cv}
\title{Fitted model validation and cross-validation (and related plots).}
\usage{
karma.cv(karma0, y = c(), test_pct = "auto", test_type = "auto",
  log = F, metric = "MAPE", xreg = NULL, fixed = F, plot = T,
  ensemble = F, cv = "out", stdout = F)
}
\arguments{
\item{karma0}{Object of class "karma.fit" or "ARIMA".}

\item{y}{A univariate time-series vector; type <numeric> or <ts>.}

\item{test_pct}{Percentage of train-test split in cross-validation (e.g. 70-30), positive integer for "window" or "percentage" test_type; "auto" to read from karma.fit object or generate; negative integer value to set window size to a multiple of the series' frequency.}

\item{test_type}{Train-test split type, i.e. percentage or fixed window; "auto": will try to read from karma.fit object or generate; "percentage": test_pct = 12 will be read as the 12 percent of the length of the series; "window": test_pct = 12 will be read as the 12 last time points (e.g. months) of the series; "auto" if input series is a ts() object, test_type is set to "window" and test_pct is set to twice the frequency of the series - if test_pct is given a negative factor, then test_pct (window size) will be set to the frequency of the series times the absolute value of that negative number.}

\item{metric}{Choose a model validation metric that will be used as the main optimisation criterion during model selection.}

\item{xreg}{Optional vector or matrix of exogenous regressors; see documentation for Arima(), package 'forecast'.}

\item{fixed}{Fixed term flag. Indicate whether the fixed term option in Arima() needs to be switched on during model selection; {T, F}; type <logical>.}

\item{plot}{Option to depict plots during local search; if TRUE (default), AC and PAC plots are active. <logical>}

\item{cv}{Choose cross-validation dataset to be used during model selection; "out": Performance of out-of-sample forecast (classic train/test split) will be used for model validation; "in": Performance of in-sample forecast (classic parametric regression type of validation) will be used for model validation.}

\item{stdout}{Option to output optimisation diagnostics during local search; <logical>}
}
\value{
Object of class "karma.fit"; (extends class "Arima" from package 'forecast').
}
\description{
Fitted model validation and cross-validation (and related plots).
}
\examples{
# Compatible with "forecast" package objects: Arima, auto.arima, nnetar, ets, bats, tbats, baggedETS, HoltWinters
# Using auto.karma():
karma.cv(auto.karma(y), cv="in")  # in-sample forecast
karma.cv(auto.karma(y), cv="out")    # out-of-sample forecast
# Using karma.fit()
karma.cv(karma.fit(y, order=c(2,1,4))); 
# Using karma.fit() with fixed terms:
karma.cv( karma.fit(y, order=list(c(1,2), 0, c(3,4)), fixed=T) ); 
# Using auto.arima():
karma.cv(auto.arima(y));
#' # Using nnetar():
karma.cv(nnetar(y));
#' # Using ets():
karma.cv(ets(y));
}
\seealso{
\code{\link{tseries}}, \code{\link{forecast}}
}

