% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karma_fit.r
\name{karma.fit}
\alias{karma.fit}
\title{Fit ARIMA model to univariate time-series. Mostly a wrapper to 'Arima' function in the 'forecast' package; the difference is that it facilitates fitting fixed term models.}
\usage{
karma.fit(y, order = c(0, 0, 0), seasonal = c(0, 0, 0), log = F,
  xreg = NULL, fixed = F, stdout = F)
}
\arguments{
\item{y}{A univariate time-series vector; type <numeric> or <ts>.}

\item{order}{Specification of the non-seasonal part of the ARIMA model (see documentation for Arima(), argument 'order').}

\item{log}{Logarithmic transformation flag; Indicates whether the input series needs to be log-transformed for stationarity; {T, F}; type <logical>.}

\item{xreg}{Optional vector or matrix of exogenous regressors; see documentation for Arima(), package 'forecast'.}

\item{fixed}{Fixed term flag. Indicate whether the fixed term option in Arima() needs to be switched on during model selection; {T, F}; type <logical>.}

\item{stdout}{Option to output optimisation diagnostics during local search; <logical>}
}
\value{
Object of class "karma.fit"; (extends class "ARIMA" from package 'forecast').
}
\description{
Fit ARIMA model to univariate time-series. Mostly a wrapper to 'Arima' function in the 'forecast' package; the difference is that it facilitates fitting fixed term models.
}
\examples{
# Fit and summarise model:
summary(karma.fit(WWWusage, order=list(c(1,2), 1, c(3,4)), fixed=T))
# Using forecast() from package 'forecast':
plot(forecast( karma.fit(WWWusage, order=c(2,1,3), fixed=F) )) 
}
\seealso{
\code{\link{tseries}}, \code{\link{forecast}}
}

