% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karma_forecast.r
\name{karma.forecast}
\alias{karma.forecast}
\title{#Forecast and plot future periods: Combine multiple weak learners (fitted models) from a pre-trained "karma.ensemble" object to produce future time periods.}
\usage{
karma.forecast(kensemble, h = 12, plot = T, xreg = NULL, stdout = F)
}
\arguments{
\item{kensemble}{Object of class "karma.ensemble"; ensemble of unique trained models. <karma.ensemble>}

\item{h}{Number of periods to forecast ahead. <numeric> integer}

\item{plot}{Option to depict plots during local search; if TRUE (default), AC and PAC plots are active. <logical>}

\item{stdout}{Option to output optimisation diagnostics during local search; <logical>}
}
\value{
Object of class "karma.forecast".
}
\description{
#Forecast and plot future periods: Combine multiple weak learners (fitted models) from a pre-trained "karma.ensemble" object to produce future time periods.
}
\examples{
# Create and forecast ensemble of 10 weak learners:
karma.forecast( karma.ensemble(JohnsonJohnson) )
}
\seealso{
\code{\link{tseries}}, \code{\link{forecast}}
}

