% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karma_portmanteau.r
\name{karma.portmanteau}
\alias{karma.portmanteau}
\title{Portmanteau test to detect statistically significant autocorrelation and partial autocorrelation lags.}
\usage{
karma.portmanteau(y, plot = F, N = 100)
}
\arguments{
\item{y}{A univariate time-series vector; type <numeric> or <ts>.}

\item{plot}{Option to depict plots during local search; if TRUE (default), AC and PAC plots are active. <logical>}

\item{N}{Maximum lag at which to calculate autocorrelation and partial autocorrelatin functions; see documentation for acf(), pacf().}
}
\value{
Object of class "karma.portmanteau".
}
\description{
Portmanteau test to detect statistically significant autocorrelation and partial autocorrelation lags.
}
\examples{
autocorrelation.lags = karma.portmanteau( magic.karma(WWWusage)$residuals )
}
\seealso{
\code{\link{tseries}}, \code{\link{forecast}}, \code{\link{car}}
}

