% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karma_transform.r
\name{karma.transform}
\alias{karma.transform}
\title{Find transformation steps required for stationarity.}
\usage{
karma.transform(yt, stdout = T, autolog = T, autodiffs = 1,
  autolags = F, r2_criterion = T)
}
\arguments{
\item{yt}{A univariate time-series vector; type <numeric> or <ts>.}

\item{stdout}{Option to print out all search diagnostics; <logical>.}

\item{autolog}{Logarithmic search flag. Indicates whether log-transformations on the input series will be part of the search.}

\item{autodiffs}{Differencing search flag. Indicates whether differencing on the input series will be part of the search.}

\item{autolags}{Flag T/F indicating whether or not to set lags automatically as a function of the length of the series.}

\item{r2_criterion}{Flag T/F incidating whether or not to use adjusted R-square as an ADF model selection criterion. When FALSE, the simplest possible stationarity transformation will be preferred.}
}
\value{
Object of class "karma.transform".
}
\description{
Find transformation steps required for stationarity.
}
\examples{
# Apply ADF test and print out diagnostics:
stationarity.options <- karma.transform(WWWusage, stdout = F, autolog = F, autodiffs = 1)
print(stationarity.options$model)  # best fitted model type
print(stationarity.options$diffs)  # best degree of differencing
print(stationarity.options$log)    # best use of logarithmic transformation 
}
\seealso{
\code{\link{tseries}}, \code{\link{forecast}}
}

