% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_boxjen.r
\name{ModelEvaluationFunction}
\alias{ModelEvaluationFunction}
\title{Automated ARIMA model selection according to flexible optimisation criteria.}
\usage{
ModelEvaluationFunction(sol, y, diffs_ = 0, metric = "MAPE", fixed_ = T,
  xreg = NULL, cv_ = "out", test_pct = 20, test_type = "percentage",
  plot = T)
}
\arguments{
\item{sol}{ARIMA order in the form of a candidate solution to the optimisation problem.}

\item{y}{A univariate time-series vector; type <numeric> or <ts>.}

\item{diffs_}{Differencing step: Indicates whether the candidate solution needs to be differenced for stationarity (and to what degree); {0,1,...,n}; type <int>.}

\item{metric}{Choose a model validation metric that will be used as the main optimisation criterion during model selection.}

\item{fixed_}{Fixed term flag. Indicate whether the fixed term option in Arima() needs to be switched on during model selection; {T, F}; type <logical>.}

\item{xreg}{Optional vector or matrix of exogenous regressors; see documentation for Arima(), package 'forecast'.}

\item{cv_}{Choose cross-validation dataset to be used during model selection; "out": Performance of out-of-sample forecast (classic train/test split) will be used for model validation; "in": Performance of in-sample forecast (classic parametric regression type of validation) will be used for model validation.}

\item{test_pct}{Percentage of train-test split in cross-validation (e.g. 70-30).}

\item{test_type}{Train-test split type, i.e. percentage or fixed window; "percentage": test_pct = 12 will be read as the 12 percent of the length of the series; "window": test_pct = 12 will be read as the 12 last time points (e.g. months) of the series.}

\item{ar_terms}{Autoregressive terms; can be in the form of a vector (fixed_ = FALSE) or a list of vectors (fixed_ = TRUE).}

\item{ma_terms}{Move average terms; can be in the form of a vector (fixed_ = FALSE) or a list of vectors (fixed_ = TRUE).}
}
\value{
Object of class "karma.fit"; (extends class "Arima" from package 'forecast').
}
\description{
Automated ARIMA model selection according to flexible optimisation criteria.
}
\examples{
sol = list(ar_terms=c(1,2), ma_terms=c(3,4)) 
ModelEvaluationFunction(y, sol, 0)
}
\seealso{
\code{\link{tseries}}, \code{\link{forecast}}
}

